+++ CMakeLists.txt
@@ -18,8 +18,9 @@ set(LOADABLE_EXTENSION_NAME ${TARGET_NAME}_loadable_extension)
 project(${TARGET_NAME})
 include_directories(src/include)
 
-set(EXTENSION_SOURCES src/ipc/array_stream.cpp src/ipc/stream_factory.cpp src/ipc/ipc_file_stream_reader.cpp
-        src/ipc/ipc_buffer_stream_reader.cpp src/scanner/read_arrow.cpp src/scanner/scan_arrow_ipc.cpp
+set(EXTENSION_SOURCES src/ipc/array_stream.cpp src/ipc/stream_factory.cpp
+        src/ipc/stream_reader/stream_reader.cpp src/ipc/stream_reader/ipc_file_stream_reader.cpp
+        src/ipc/stream_reader/ipc_buffer_stream_reader.cpp src/scanner/read_arrow.cpp src/scanner/scan_arrow_ipc.cpp
         src/nanoarrow_extension.cpp src/write_arrow_stream.cpp)
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES})
