+++ src/include/ipc/stream_reader/stream_reader.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB - NanoArrow
 //
-// ipc/stream_reader.hpp
+// ipc/stream_reader/stream_reader.hpp
 //
 //
 //===----------------------------------------------------------------------===//
@@ -87,38 +87,5 @@ protected:
 
 };
 
-//! Buffer Stream
-class IPCBufferStreamReader final : public IPCStreamReader {
-public:
-  IPCBufferStreamReader(vector<ArrowIPCBuffer> buffers, Allocator& allocator);
-
-  ArrowIpcMessageType ReadNextMessage() override;
-private:
-  vector<ArrowIPCBuffer> buffers;
-  idx_t cur_idx = 0;
-};
-
-//! IPC File
-class IPCFileStreamReader final : public IPCStreamReader {
- public:
-  IPCFileStreamReader(FileSystem& fs, unique_ptr<FileHandle> handle, Allocator& allocator);
-
-  ArrowIpcMessageType ReadNextMessage() override;
-
- private:
-  static constexpr uint32_t kContinuationToken = 0xFFFFFFFF;
-
-  BufferedFileReader file_reader;
-
-  ArrowIpcMessagePrefix message_prefix{};
-  AllocatedData message_header;
-  shared_ptr<AllocatedData> message_body;
-
-  void EnsureInputStreamAligned();
-
-  static void DecodeArray(nanoarrow::ipc::UniqueDecoder &decoder, ArrowArray* out,  ArrowBufferView& body_view, ArrowError *error);
-
-  void PopulateNames(vector<string>& names);
-};
 } // namespace ext_nanoarrow
 } // namespace duckdb
