+++ src/ipc/stream_factory.cpp
@@ -1,4 +1,6 @@
 #include "ipc/stream_factory.hpp"
+#include "ipc/stream_reader/ipc_buffer_stream_reader.hpp"
+#include "ipc/stream_reader/ipc_file_stream_reader.hpp"
 
 namespace duckdb {
 namespace ext_nanoarrow {
