+++ src/scanner/read_arrow.cpp
@@ -19,7 +19,7 @@
 
 #include "nanoarrow_errors.hpp"
 #include "ipc/stream_factory.hpp"
-#include "ipc/stream_reader.hpp"
+#include "ipc/stream_reader/stream_reader.hpp"
 #include "table_function/arrow_ipc_function_data.hpp"
 
 // read_arrow() implementation
